:recycle: Refactoring code.  数据权限 DataScope.java 增加链式操作方便使用

+++ pigx-common/pigx-common-data/src/main/java/com/pig4cloud/pigx/common/data/datascope/SelectCountByScope.java
@@ -27,7 +27,7 @@ public class SelectCountByScope extends AbstractMethod {
 				this.sqlComment());
 		SqlSource sqlSource = languageDriver.createSqlSource(configuration, sql, modelClass);
 
-		return this.addSelectMappedStatementForOther(mapperClass, sqlSource, Integer.class);
+		return this.addSelectMappedStatementForOther(mapperClass, sqlSource, Long.class);
 	}
 
 }
